+++ app/_data-models/Upscalers.ts
@@ -1,5 +1,6 @@
 class Upscalers {
   static upscalers = [
+    'RealESRGAN_x2plus',
     'RealESRGAN_x4plus',
     'RealESRGAN_x4plus_anime_6B',
     'NMKD_Siax',